+++ packages/enoki/src/index.ts
@@ -13,7 +13,10 @@ export {
 export { createDefaultEncryption, type Encryption } from './encryption.js';
 export { EnokiKeypair, EnokiPublicKey } from './EnokiKeypair.js';
 
-export { registerEnokiWallets, type EnokiWallet, enokiWalletsInitializer } from './wallet/index.js';
+export type { EnokiWallet } from './wallet/wallet.js';
+export { registerEnokiWallets } from './wallet/register.js';
+export { enokiWalletsInitializer } from './wallet/initializer.js';
+
 export {
 	isEnokiWallet,
 	isGoogleWallet,
